+++ src/sort/quick_sort.c
@@ -1,8 +1,8 @@
 #include "sort.h"
 
-static void swap(ssize_t* a, ssize_t* b)
+static void swap(int64_t* a, int64_t* b)
 {
-	ssize_t t = *a;
+	int64_t t = *a;
 	*a = *b;
 	*b = t;
 }
@@ -12,9 +12,9 @@ the pivot element at its correct position in sorted
 array, and places all smaller (smaller than pivot)
 to left of pivot and all greater elements to right
 of pivot */
-static int partition(ssize_t arr[], ssize_t low, ssize_t high)
+static int partition(int64_t arr[], int64_t low, int64_t high)
 {
-	ssize_t pivot = arr[high]; // pivot
+	int64_t pivot = arr[high]; // pivot
 	int i = (low - 1); // Index of smaller element and indicates the right position of pivot found so far
 
 	while (low <= high - 1)
@@ -35,13 +35,13 @@ static int partition(ssize_t arr[], ssize_t low, ssize_t high)
 arr[] --> Array to be sorted,
 low --> Starting index,
 high --> Ending index */
-void quick_sort(ssize_t *arr, int low, int high)
+void quick_sort(int64_t *arr, int low, int high)
 {
 	if (low < high)
 	{
 		/* pi is partitioning index, arr[pi] is now
 		at right place */
-		ssize_t pi = partition(arr, low, high);
+		int64_t pi = partition(arr, low, high);
 
 		// Separately sort elements before
 		// partition and after partition
